requirements.txt
whisperx
torch
torchaudio
pyannote.audio
